+++ plugins/destination/clickhouse/client/migrate.go
@@ -4,24 +4,23 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/apache/arrow/go/v13/arrow"
 	"github.com/cloudquery/cloudquery/plugins/destination/clickhouse/queries"
 	"github.com/cloudquery/cloudquery/plugins/destination/clickhouse/typeconv"
 	"github.com/cloudquery/cloudquery/plugins/destination/clickhouse/util"
 	"github.com/cloudquery/plugin-pb-go/specs"
-	"github.com/cloudquery/plugin-sdk/v2/schema"
+	"github.com/cloudquery/plugin-sdk/v3/schema"
 	"golang.org/x/exp/slices"
 	"golang.org/x/sync/errgroup"
 )
 
 // Migrate relies on the CLI/client to lock before running migration.
-func (c *Client) Migrate(ctx context.Context, scs schema.Schemas) error {
-	have, err := c.getTableDefinitions(ctx, scs)
+func (c *Client) Migrate(ctx context.Context, tables schema.Tables) error {
+	have, err := c.getTableDefinitions(ctx, tables)
 	if err != nil {
 		return err
 	}
 
-	want, err := typeconv.CanonizedSchemas(scs)
+	want, err := typeconv.CanonizedTables(tables)
 	if err != nil {
 		return err
 	}
@@ -39,17 +38,16 @@ func (c *Client) Migrate(ctx context.Context, scs schema.Schemas) error {
 	for _, want := range want {
 		want := want
 		eg.Go(func() (err error) {
-			tableName := schema.TableName(want)
-			c.logger.Info().Str("table", tableName).Msg("Migrating table started")
+			c.logger.Info().Str("table", want.Name).Msg("Migrating table started")
 			defer func() {
-				c.logger.Err(err).Str("table", tableName).Msg("Migrating table done")
+				c.logger.Err(err).Str("table", want.Name).Msg("Migrating table done")
 			}()
-			if len(want.Fields()) == 0 {
-				c.logger.Warn().Str("table", tableName).Msg("Table with no columns, skip")
+			if len(want.Columns) == 0 {
+				c.logger.Warn().Str("table", want.Name).Msg("Table with no columns, skip")
 				return nil
 			}
 
-			have := have.SchemaByName(tableName)
+			have := have.Get(want.Name)
 			if have == nil {
 				return c.createTable(ctx, want)
 			}
@@ -61,23 +59,23 @@ func (c *Client) Migrate(ctx context.Context, scs schema.Schemas) error {
 	return eg.Wait()
 }
 
-func unsafeSchemaChanges(have, want schema.Schemas) map[string]schema.FieldChanges {
-	result := make(map[string]schema.FieldChanges)
+func unsafeSchemaChanges(have, want schema.Tables) map[string][]schema.TableColumnChange {
+	result := make(map[string][]schema.TableColumnChange)
 	for _, w := range want {
-		current := have.SchemaByName(schema.TableName(w))
+		current := have.Get(w.Name)
 		if current == nil {
 			continue
 		}
-		unsafe := unsafeChanges(schema.GetSchemaChanges(w, current))
+		unsafe := unsafeChanges(w.GetChanges(current))
 		if len(unsafe) > 0 {
-			result[schema.TableName(w)] = unsafe
+			result[w.Name] = unsafe
 		}
 	}
 	return result
 }
 
-func unsafeChanges(changes []schema.FieldChange) schema.FieldChanges {
-	unsafe := make([]schema.FieldChange, 0, len(changes))
+func unsafeChanges(changes []schema.TableColumnChange) []schema.TableColumnChange {
+	unsafe := make([]schema.TableColumnChange, 0, len(changes))
 	for _, c := range changes {
 		if needsTableDrop(c) {
 			unsafe = append(unsafe, c)
@@ -86,10 +84,10 @@ func unsafeChanges(changes []schema.FieldChange) schema.FieldChanges {
 	return slices.Clip(unsafe)
 }
 
-func (c *Client) createTable(ctx context.Context, sc *arrow.Schema) (err error) {
-	c.logger.Debug().Str("table", schema.TableName(sc)).Msg("Table doesn't exist, creating")
+func (c *Client) createTable(ctx context.Context, table *schema.Table) (err error) {
+	c.logger.Debug().Str("table", table.Name).Msg("Table doesn't exist, creating")
 
-	query, err := queries.CreateTable(sc, c.spec.Cluster, c.spec.Engine)
+	query, err := queries.CreateTable(table, c.spec.Cluster, c.spec.Engine)
 	if err != nil {
 		return err
 	}
@@ -100,20 +98,20 @@ func (c *Client) createTable(ctx context.Context, sc *arrow.Schema) (err error) 
 	return nil
 }
 
-func (c *Client) dropTable(ctx context.Context, sc *arrow.Schema) error {
-	c.logger.Debug().Str("table", schema.TableName(sc)).Msg("Dropping table")
+func (c *Client) dropTable(ctx context.Context, table *schema.Table) error {
+	c.logger.Debug().Str("table", table.Name).Msg("Dropping table")
 
-	return c.conn.Exec(ctx, queries.DropTable(sc, c.spec.Cluster))
+	return c.conn.Exec(ctx, queries.DropTable(table, c.spec.Cluster))
 }
 
-func needsTableDrop(change schema.FieldChange) bool {
+func needsTableDrop(change schema.TableColumnChange) bool {
 	// We can safely add a nullable column without dropping the table
-	if change.Type == schema.TableColumnChangeTypeAdd && change.Current.Nullable {
+	if change.Type == schema.TableColumnChangeTypeAdd && !change.Current.NotNull {
 		return false
 	}
 
 	// We can safely ignore removal of nullable columns without dropping the table
-	if change.Type == schema.TableColumnChangeTypeRemove && change.Previous.Nullable {
+	if change.Type == schema.TableColumnChangeTypeRemove && !change.Previous.NotNull {
 		return false
 	}
 
@@ -121,8 +119,8 @@ func needsTableDrop(change schema.FieldChange) bool {
 	return true
 }
 
-func (c *Client) autoMigrate(ctx context.Context, have, want *arrow.Schema) error {
-	changes := schema.GetSchemaChanges(want, have)
+func (c *Client) autoMigrate(ctx context.Context, have, want *schema.Table) error {
+	changes := want.GetChanges(have)
 
 	if unsafe := unsafeChanges(changes); len(unsafe) > 0 {
 		// we can get here only with migrate_mode: forced
@@ -133,16 +131,15 @@ func (c *Client) autoMigrate(ctx context.Context, have, want *arrow.Schema) erro
 		return c.createTable(ctx, want)
 	}
 
-	tableName := schema.TableName(want)
 	for _, change := range changes {
 		// we only handle new columns
 		if change.Type != schema.TableColumnChangeTypeAdd {
 			continue
 		}
 
-		c.logger.Debug().Str("table", tableName).Str("column", change.Current.Name).Msg("Adding new column")
+		c.logger.Debug().Str("table", want.Name).Str("column", change.Current.Name).Msg("Adding new column")
 
-		query, err := queries.AddColumn(tableName, c.spec.Cluster, change.Current)
+		query, err := queries.AddColumn(want.Name, c.spec.Cluster, change.Current)
 		if err != nil {
 			return err
 		}