+++ plugins/destination/gremlin/client/deletestale.go
@@ -5,10 +5,10 @@ import (
 	"time"
 
 	gremlingo "github.com/apache/tinkerpop/gremlin-go/v3/driver"
-	"github.com/cloudquery/plugin-sdk/v2/schema"
+	"github.com/cloudquery/plugin-sdk/v3/schema"
 )
 
-func (c *Client) DeleteStale(ctx context.Context, tables schema.Schemas, source string, syncTime time.Time) error {
+func (c *Client) DeleteStale(ctx context.Context, tables schema.Tables, source string, syncTime time.Time) error {
 	session, closer, err := c.newSession()
 	if err != nil {
 		return err
@@ -16,10 +16,9 @@ func (c *Client) DeleteStale(ctx context.Context, tables schema.Schemas, source 
 	defer closer()
 
 	for _, table := range tables {
-		tableName := schema.TableName(table)
 		g := gremlingo.Traversal_().WithRemote(session).
 			V().
-			HasLabel(tableName).
+			HasLabel(table.Name).
 			Has(schema.CqSourceNameColumn.Name, source).
 			Has(schema.CqSyncTimeColumn.Name, gremlingo.P.Lt(syncTime)).
 			SideEffect(AnonT.Drop())